refactor entity and relationship set types into separate modules (#616)
